Create Adding a method that takes parameters in a class
As you can see, the setDim() method is used to set the dimensions of each box.
For example, when
mybox1.setDim(10,20,15);
is executed,10 is copied into parameter w, 20 is copied to parameter h, 15 is copied to parameter d. Inside setDim() the values of w, h and d are then assigned to width ,height and depth respectively.